Moved non-whitebox test for "search" pkg to blackbox test
