+++ search/search_repository_blackbox_test.go
@@ -3,14 +3,19 @@ package search_test
 import (
 	"context"
 	"fmt"
+	"math"
+	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/fabric8-services/fabric8-wit/gormtestsupport"
+	"github.com/fabric8-services/fabric8-wit/rendering"
 	"github.com/fabric8-services/fabric8-wit/resource"
 	"github.com/fabric8-services/fabric8-wit/search"
 	tf "github.com/fabric8-services/fabric8-wit/test/testfixture"
 	"github.com/fabric8-services/fabric8-wit/workitem"
 	"github.com/fabric8-services/fabric8-wit/workitem/link"
+	_ "github.com/lib/pq"
 	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -281,5 +286,272 @@ func containsAllWorkItems(expectedWorkitems []workitem.WorkItem, actualWorkitems
 			}
 		}
 		return true
+	}
+}
+
+func (s *searchRepositoryBlackboxTest) TestSearch() {
+
+	var start, limit int = 0, 100
+
+	s.T().Run("Search accross title and description", func(t *testing.T) {
+
+		t.Run("match title and descrition", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "test sbose title '12345678asdfgh'"
+					fxt.WorkItems[idx].Fields[workitem.SystemDescription] = rendering.NewMarkupContentFromLegacy(`"description" for search test`)
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `Sbose "deScription" '12345678asdfgh'`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 1)
+		})
+
+		t.Run("match title with description undefined", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "test nofield sbose title '12345678asdfgh'"
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `sbose nofield`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 1)
+		})
+
+		t.Run("match title with slash", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "add new error types in models/errors.go'"
+					fxt.WorkItems[idx].Fields[workitem.SystemDescription] = rendering.NewMarkupContentFromLegacy(`Make sure remoteworkitem can access..`)
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `models/errors.go remoteworkitem`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 1)
+		})
+
+		t.Run("match title with braces", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "Bug reported by administrator for input = (value)"
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `(value)`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 1)
+
+		})
+
+		t.Run("match title with braces and brackets", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "trial for braces (pranav) {shoubhik} [aslak]"
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `(pranav) {shoubhik} [aslak]`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 1)
+
+		})
+
+		t.Run("no match", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "test should return 0 results'"
+				}
+				return nil
+			}))
+			// when
+			searchQuery := `negative case`
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			verify(t, searchQuery, searchResults, 0)
+		})
+	})
+
+	s.T().Run("Search by number", func(t *testing.T) {
+		fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10))
+		spaceID := fxt.Spaces[0].ID.String()
+
+		t.Run("and by space", func(t *testing.T) {
+			t.Run("single match", func(t *testing.T) {
+				// given
+				queryNumber := fxt.WorkItems[2].Number
+				// when looking for `number:3`
+				searchQuery := fmt.Sprintf("number:%d", queryNumber)
+				searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+				// then there should be a single match
+				require.NoError(t, err)
+				require.Len(t, searchResults, 1)
+				assert.Equal(t, queryNumber, searchResults[0].Number)
+			})
+
+			t.Run("multiple matches", func(t *testing.T) {
+				// given
+				queryNumber := fxt.WorkItems[0].Number
+				// when looking for `number:1`
+				searchQuery := fmt.Sprintf("number:%d", queryNumber)
+				searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+				// then there should be 2 matches: `1` and `10`
+				require.NoError(t, err)
+				require.Len(t, searchResults, 2)
+				for _, searchResult := range searchResults {
+					// verifies that the number in the search result contains the query number
+					assert.Contains(t, strconv.Itoa(searchResult.Number), strconv.Itoa(queryNumber))
+				}
+			})
+			t.Run("not found", func(t *testing.T) {
+				// given
+				notExistingWINumber := 12345 // We only created one work item in that space, so that number should not exist
+				searchString := "number:" + strconv.Itoa(notExistingWINumber)
+				// when
+				workItemList, _, err := s.searchRepo.SearchFullText(context.Background(), searchString, &start, &limit, &spaceID)
+				// then
+				require.NoError(t, err)
+				require.Len(t, workItemList, 0)
+			})
+		})
+		t.Run("not by space", func(t *testing.T) {
+			t.Run("single match", func(t *testing.T) {
+				// given
+				searchString := "number:" + strconv.Itoa(fxt.WorkItems[0].Number)
+				// when
+				workItemList, _, err := s.searchRepo.SearchFullText(context.Background(), searchString, &start, &limit, nil)
+				// then
+				require.NoError(t, err)
+				require.True(t, len(workItemList) >= 1, "at least one work item should be found for the given work item number")
+				var found bool
+				for _, wi := range workItemList {
+					if wi.ID == fxt.WorkItems[0].ID {
+						found = true
+					}
+				}
+				require.True(t, found, "failed to found: %s", fxt.WorkItems[0].ID)
+			})
+			t.Run("not found", func(t *testing.T) {
+				// given
+				notExistingWINumber := math.MaxInt64 - 1 // That ID most likely does not exist at all
+				searchString := "number:" + strconv.Itoa(notExistingWINumber)
+				// when
+				workItemList, _, err := s.searchRepo.SearchFullText(context.Background(), searchString, &start, &limit, nil)
+				// then
+				require.NoError(t, err)
+				require.Len(t, workItemList, 0)
+			})
+		})
+	})
+
+	s.T().Run("Search by URL - single match", func(t *testing.T) {
+		t.Run("single match", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "test nofield sbose title '12345678asdfgh'"
+				}
+				return nil
+			}))
+			// when looking for `http://.../3` there should be a single match
+			queryNumber := fxt.WorkItems[2].Number
+			searchQuery := fmt.Sprintf("%s%d", "http://demo.openshift.io/work-item/list/detail/", queryNumber)
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			require.Len(t, searchResults, 1)
+			assert.Equal(t, queryNumber, searchResults[0].Number)
+		})
+
+		t.Run("multiple matches", func(t *testing.T) {
+			// given
+			fxt := tf.NewTestFixture(t, s.DB, tf.WorkItems(10, func(fxt *tf.TestFixture, idx int) error {
+				if idx == 0 {
+					fxt.WorkItems[idx].Fields[workitem.SystemTitle] = "test nofield sbose title '12345678asdfgh'"
+				}
+				return nil
+			}))
+			// when looking for `http://.../1` there should be a 2 matchs: `http://.../1` and `http://.../10``
+			queryNumber := fxt.WorkItems[0].Number
+			searchQuery := fmt.Sprintf("%s%d", "http://demo.openshift.io/work-item/list/detail/", queryNumber)
+			spaceID := fxt.Spaces[0].ID.String()
+			searchResults, _, err := s.searchRepo.SearchFullText(context.Background(), searchQuery, &start, &limit, &spaceID)
+			// then
+			require.NoError(t, err)
+			require.Len(t, searchResults, 2)
+			for _, searchResult := range searchResults {
+				// verifies that the number in the search result contains the query number
+				assert.Contains(t, strconv.Itoa(searchResult.Number), strconv.Itoa(queryNumber))
+			}
+		})
+	})
+}
+
+// verify verifies that the search results match with the expected count and that the title or description contain all
+// the terms of the search query
+func verify(t *testing.T, searchQuery string, searchResults []workitem.WorkItem, expectedCount int) {
+	// Since this test adds test data, whether or not other workitems exist
+	// there must be at least 1 search result returned.
+	if len(searchResults) == expectedCount && expectedCount == 0 {
+		// no point checking further, we got what we wanted.
+		return
+	}
+	require.Equal(t, expectedCount, len(searchResults), "invalid number of results in the search")
+
+	// These keywords need a match in the textual part.
+	allKeywords := strings.Fields(searchQuery)
+	// These keywords need a match optionally either as URL string or ID		 +				keyWord = strings.ToLower(keyWord)
+	// optionalKeywords := []string{workItemURLInSearchString, strconv.Itoa(fxt.WorkItems[idx].Number)}
+	// We will now check the legitimacy of the search results.
+	// Iterate through all search results and see whether they meet the criteria
+	for _, searchResult := range searchResults {
+		t.Logf("Examining workitem id=`%v` number=`%d` using keywords %v", searchResult.ID, searchResult.Number, allKeywords)
+		for _, keyWord := range allKeywords {
+			keyWord = strings.ToLower(keyWord)
+			t.Logf("Verifying workitem id=`%v` number=`%d` for keyword `%s`...", searchResult.ID, searchResult.Number, keyWord)
+			workItemTitle := ""
+			if searchResult.Fields[workitem.SystemTitle] != nil {
+				workItemTitle = strings.ToLower(searchResult.Fields[workitem.SystemTitle].(string))
+			}
+			workItemDescription := ""
+			if searchResult.Fields[workitem.SystemDescription] != nil {
+				descriptionField := searchResult.Fields[workitem.SystemDescription].(rendering.MarkupContent)
+				workItemDescription = strings.ToLower(descriptionField.Content)
+			}
+			assert.True(t,
+				strings.Contains(workItemTitle, keyWord) || strings.Contains(workItemDescription, keyWord),
+				"`%s` neither found in title `%s` nor in the description `%s` for workitem #%d", keyWord, workItemTitle, workItemDescription, searchResult.Number)
+		}
 	}
 }
